+++ spec/dummy_app/app/models/wotsit.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 class Wotsit < ApplicationRecord
+  attr_readonly :id
   has_paper_trail
 
   belongs_to :widget, optional: true