+++ CHANGELOG.md
@@ -17,6 +17,13 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 - None
 
+### Dependencies
+
+- Drop support for Rails 6.1, which [reached EoL on 2024-10-01][2]
+- Drop support for Rails 7.0, which [reached EoL on 2025-04-01][2]
+- Drop support for Ruby 3.0, which [reached EoL on 2024-04-23][3]
+- Drop support for Ruby 3.1, which [reached EoL on 2024-03-31][3]
+
 ## 16.0.0 (2024-11-08)
 
 ### Breaking Changes