+++ CHANGELOG.md
@@ -11,7 +11,9 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Added
 
-- None
+- [#1459](https://github.com/paper-trail-gem/paper_trail/issues/1459) 
+  `rails generate paper_trail:install` now creates a `paper_trail.rb` initializer
+  that disables versioning on touch events by default.
 
 ### Fixed