+++ CHANGELOG.md
@@ -11,7 +11,7 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Added
 
-- None
+- `rails generate paper_trail:install` now accepts an argument for custom versions table, e.g. `rails generate paper_trail:install CommentVersion` created `comment_versions` table
 
 ### Fixed