+++ CHANGELOG.md
@@ -11,7 +11,10 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Added
 
-- None
+- `rails generate paper_trail:install` now accepts an argument for custom versions table, e.g.
+  `rails generate paper_trail:install CommentVersion` created `comment_versions` table
+- `rails generate paper_trail:update_item_subtype` now supports custom version classes via 
+  `--version-class-name` option, e.g. `--version-class-name=CommentVersion`
 
 ### Fixed